requirements.txt
django-tinymce
fontawesomefree
django-crispy-forms
django-recaptcha
pillow
six